Implements Logging in Bolt Application
This change integrates logging in the application, so standard tasks and extensions can use the default logging mechanism to report different actions.

I also refactored the bolt module and put most of the code in a _BoltApplication class that it gets initialized when the module is loaded. This makes a lot easier to share state without using global variables.

Implements: story/issue #19