Implementation of Delete File Task
Description
-----------
This change implements the delete file task, which searches for files matching the configuration options and deletes them.

Testing Performed
-----------------
Unit tests have been provided for all the functionality added. I also ran the task through the application to delete all the `.pyc` files and it worked as expected.