Enhances Spawn Task to Terminate Process and Children (#127)
## Summary
This change modifies the `SpawnExecuteTask` to leverage `psutil`, if
available, to not only terminate the spawn process at the end, but also
all the child processes. If `psutil` is not available, the task resorts
to the previous behavior and only kills the spawn process. This may
result on dangling process in some instances.

One of the reasons for this change is to be able to launch a web
application behind a server using `npm`. NPM is notorious for leaving
child processes behind if they are not explicitly killed. This
implementation, provided `psutil` is installed, will ensure all
processes are terminated.